requirements.txt
python-telegram-bot~=22.1
PyYAML~=6.0.2
requests~=2.32.4
bs4~=0.0.2
beautifulsoup4~=4.13.4
selenium~=4.33.0